スクール予約機能

+++ app/Http/Controllers/Authenticated/Calendar/General/CalendarsController.php
@@ -40,7 +40,7 @@ class CalendarsController extends Controller
 
     public function delete(Request $request){
         $delete_id=$request->delete_date;
-        ReserveSettingUsers::where('reserve_setting_id',$delete_id)->delete();
+        ReserveSettingUsers::where('user_id',Auth::id())->where('reserve_setting_id',$delete_id)->delete();
         $delete_reserve=ReserveSettings::where('id',$delete_id)->first();
         $delete_reserve->increment('limit_users');
